requirements.txt
python-telegram-bot==13.7
yt-dlp==2023.11.16
requests==2.31.0